+++ detectDuplicateAccounts.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python -m datacrawler.détecte_doublons_comptes